pyproject.toml
[tool.poetry]
name = "argo-workflow-runner"
version = "0.1.0"
description = "A Python runner for Argo Workflows"
authors = ["Sam Gao <samuel.gao023@gmail.com> Hunter Han <hanyh2004@gmail.com>"]
readme = "README.md"

homepage = "https://ai3labs.net"
repository = "https://github.com/yourusername/argo-workflow-runner"
documentation = "https://argo-workflow-runner.readthedocs.io"
packages = [
    { include = "argo_workflow_runner", from = "src" }
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = ">=3.10.10,<4.0"
langchain-community="^0.3.3"
langchain-openai="^0.2.14"
langgraph="^0.2.39"
fastapi = "^0.115.6"
redis = "^5.2.1"
python-dotenv = "^1.0.1"
uvicorn = "^0.34.0"
python-decouple = "^3.8"
aiofiles = "^24.1.0"
websockets = "13.1"
aiohttp = "^3.11.11"
python-multipart = "^0.0.20"
tavily-python = "^0.5.0"
python-ulid = "^3.0.0"
passlib = "^1.7.4"
httpx = "^0.28.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
black = "^22.0.0"
flake8 = "^4.0.0"
mypy = "^0.950"
pytest-cov = "^3.0.0"

[tool.poetry.scripts]
argo-runner = "argo_workflow_runner.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
